# docs/swagger.yaml
definitions:
...
  db.ErrorMessage:
    properties:
      code:
        type: integer
      messages:
        type: object
    type: object
...
info:
  contact: {}
  title: Find Your Car
  version: 0.0.9
paths:
...
  /fyc/errors:
    delete:
      description: Delete a specific language entry from the messages field of an
        error message by code
      parameters:
      - description: Error message code
        in: query
        name: code
        required: true
        type: string
      - description: Language of the error message
        in: query
        name: lang
        required: true
        type: string
      responses: {}
      summary: Delete a specific language from an error message
      tags:
      - Errors
    get:
      description: Get a list of all error messages or a specific one by code and
        language
      parameters:
      - description: Error code to fetch specific error message
        in: query
        name: code
        type: string
      - description: Language of the error message
        in: query
        name: lang
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.ErrorMessage'
            type: array
      summary: Get all error messages or a specific one
      tags:
      - Errors
    post:
      consumes:
      - application/json
      description: Create a new error message
      parameters:
      - description: Error message object
        in: body
        name: errMsg
        required: true
        schema:
          $ref: '#/definitions/db.ErrorMessage'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/db.ErrorMessage'
      summary: Create a new error message
      tags:
      - Errors
    put:
      consumes:
      - application/json
      description: Update an existing error message by code
      parameters:
      - description: Error message code
        in: query
        name: code
        required: true
        type: string
      - description: Updated error message object
        in: body
        name: errMsg
        required: true
        schema:
          $ref: '#/definitions/db.ErrorMessage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.ErrorMessage'
      summary: Update an existing error message
      tags:
      - Errors
...
securityDefinitions:
  BearerAuth3rdParty:
    description: 'Authorization token for third-party section (Ensure the token is
      in this format: Bearer token)'
    in: header
    name: Authorization
    type: apiKey
  BearerAuthBackOffice:
    description: 'Authorization token for back-office section (Ensure the token is
      in this format: Bearer token)'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"